Use intp as label dtype
numpy.bincount seems to cast its input to dtype('intp') according to
the rule safe. Therefore giving a dtype such as uint32 or int64 that
can hold a larger number fails on 32-bit platforms.

+++ skimage/morphology/misc.py
@@ -338,7 +338,7 @@ def remove_close_objects(
     )
 
     # Label objects, only samples where labels != 0 are evaluated
-    labels = np.empty_like(image, dtype=np.uint32)
+    labels = np.empty_like(image, dtype=np.intp)
     ndi.label(image, selem, output=labels)
 
     if priority is None:
